fix: always show confirmation dialog when disconnecting alby account
